Avoid moving input parameters. They are built directly into the Base now.
The previous algorithm used to create an internal representation from the user arguments.
But that required us to move from this internal representation out onto the final representation.

Besides being not efficient, it also messed up the ergonomics. For example, we could not have a std::atomic<int>
because we would receive an `int` parameter and create `std::atomic<int>` as internal representation.
Then we would try to move `std::atomic<int>` into the final `std::atomic<int>` which is not a valid operation.

Now, the intermediate representation only contains either `Ignore`s or `ArrayBuilder<T>`.
By doing so, we don't consume the input parameters.

After the calculation is done, we create the Base forwarding the arguments, and then fill the data for arrays.